Move stuff around, add a test
